fix: Enum 패키지 변경, 명명 변경 (RewardEnum -> Reward)
