+++ src/main/java/lotto/README.md
@@ -96,9 +96,9 @@
   - [ ] 당첨 결과를 구하는 책임 분리
   - [ ] 당첨 결과를 객체로 도출 -> 수익률은 당첨 결과에게 조회하도록
 - RewardEnum 
-  - [ ] 명명 변경: RewardEnum -> Reward
-  - [ ] 패키지 분리 필요 없음
-  - [ ] 큰 숫자에 _추가
+  - [x] 명명 변경: RewardEnum -> Reward
+  - [x] 패키지 분리 필요 없음
+  - [x] 큰 숫자에 _추가
   - [ ] countMap에 사용
   - [ ] matchingCount에 해당하는 타입이 없다면 null을 리턴하기 보다 미당첨 타입을 리턴
 - 모든 도메인 로직에 test 추가
